% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myfix.R
\name{mf}
\alias{mf}
\title{My String Fix Function - Shorthand}
\usage{
mf()
}
\value{
The string in your clipboard is edited to make windows paths R friendly.
}
\description{
This function converts a windows path on your clipboard to an R-friendly path on your clipboard.
}
\details{
This function uses normalizePath() to convert backslashes to forward slashes in the text on
your clipboard. It uses gsub() to remove any quotation marks, then writes the resulting text to your clipboard.

\if{html}{\out{<div class="sourceCode">}}\preformatted{mf <- function() \{
  str <- readClipboard()
  str <- gsub('"', "", str)
  str <- normalizePath(str, winslash = "/")
  str <- charToRaw(paste0(str, " "))
  writeClipboard(str)
\}
}\if{html}{\out{</div>}}
}
\examples{
mf()
}
