% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myfix.R
\name{mp}
\alias{mp}
\title{Load essential packages}
\usage{
mp()
}
\description{
Loads the packages I use every day: tidvyerse, myfix, survival, survminer, readxl, ggsci, and broom
}
\details{
\if{html}{\out{<div class="sourceCode">}}\preformatted{}\if{html}{\out{</div>}}
}
\examples{
mp()
}
