% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myfix.R
\name{pamcol}
\alias{pamcol}
\title{Load PAM50 color palette}
\usage{
pamcol()
}
\value{
A vector of four colors, representing Luminal A, Luminal B, HER2-enriched and Basal-like breast cancer.
}
\description{
Create a color palette for PAM50
}
\details{
\if{html}{\out{<div class="sourceCode">}}\preformatted{Colors: c("blue", "deepskyblue", "magenta", "red")
}\if{html}{\out{</div>}}
}
\examples{
palette <- pamcol()
}
