% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myfix.R
\name{vrn}
\alias{vrn}
\title{Viedoc rename function}
\usage{
vrn(x)
}
\value{
A list of data sheets with simplified variable names
}
\description{
This function simplifies the most common variable names in Viedoc
}
\details{
List of changes:

\if{html}{\out{<div class="sourceCode">}}\preformatted{
vrn <- function(x) \{
 sapply(x, function(y)\{
 if("SubjectId" \%in\% names(y))
 y \%>\%
    rename(id = SubjectId,
           event = EventName,
           event.date = EventDate)
 else y
 \})
\}
}\if{html}{\out{</div>}}
}
\examples{
vrn(x)
}
